services/log-consumer/requirements.txt
# services/log-consumer/requirements.txt
kafka-python
protobuf